# backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Daily Meal V1 API",
    "version": "1.0.0",
    "description": "个人饮食管理后端。业务时区默认 Asia/Shanghai。所有营养计算由 Go 完成，未知值保留 null。V1 不含登录或多用户。"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/goal-settings/default": {
      "get": {
        "summary": "当前默认目标；首次设置前为 null",
        "operationId": "get_api_v1_goal_settings_default",
        "tags": [
          "goal-settings"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Goal"
                    }
                  ],
                  "nullable": true
                },
                "example": {
                  "energy_kcal": 2000,
                  "protein_percent": 30,
                  "carbohydrate_percent": 40,
                  "fat_percent": 30,
                  "targets": {
                    "energy_kcal": 2000,
                    "protein_g": 150,
                    "carbohydrate_g": 200,
                    "fat_g": 66.66666666666667
                  },
                  "effective_on": "2026-09-08"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      },
      "put": {
        "summary": "从业务时区今天生效，保留历史版本",
        "operationId": "put_api_v1_goal_settings_default",
        "tags": [
          "goal-settings"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                },
                "example": {
                  "energy_kcal": 2000,
                  "protein_percent": 30,
                  "carbohydrate_percent": 40,
                  "fat_percent": 30,
                  "targets": {
                    "energy_kcal": 2000,
                    "protein_g": 150,
                    "carbohydrate_g": 200,
                    "fat_g": 66.66666666666667
                  },
                  "effective_on": "2026-09-08"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalInput"
              },
              "example": {
                "energy_kcal": 2000,
                "protein_percent": 30,
                "carbohydrate_percent": 40,
                "fat_percent": 30
              }
            }
          }
        }
      }
    },
    "/api/v1/daily-goals/{date}": {
      "get": {
        "summary": "解析日期目标",
        "operationId": "get_api_v1_daily_goals_date",
        "tags": [
          "daily-goals"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedGoal"
                },
                "example": {
                  "date": "2026-09-08",
                  "source": "default",
                  "goal": {
                    "energy_kcal": 2000,
                    "protein_percent": 30,
                    "carbohydrate_percent": 40,
                    "fat_percent": 30,
                    "targets": {
                      "energy_kcal": 2000,
                      "protein_g": 150,
                      "carbohydrate_g": 200,
                      "fat_g": 66.66666666666667
                    },
                    "effective_on": "2026-09-08"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2026-09-08"
            }
          }
        ]
      },
      "put": {
        "summary": "设置日期覆盖",
        "operationId": "put_api_v1_daily_goals_date",
        "tags": [
          "daily-goals"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedGoal"
                },
                "example": {
                  "date": "2026-09-08",
                  "source": "default",
                  "goal": {
                    "energy_kcal": 2000,
                    "protein_percent": 30,
                    "carbohydrate_percent": 40,
                    "fat_percent": 30,
                    "targets": {
                      "energy_kcal": 2000,
                      "protein_g": 150,
                      "carbohydrate_g": 200,
                      "fat_g": 66.66666666666667
                    },
                    "effective_on": "2026-09-08"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2026-09-08"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalInput"
              },
              "example": {
                "energy_kcal": 2000,
                "protein_percent": 30,
                "carbohydrate_percent": 40,
                "fat_percent": 30
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除覆盖，恢复该日默认（幂等）",
        "operationId": "delete_api_v1_daily_goals_date",
        "tags": [
          "daily-goals"
        ],
        "responses": {
          "204": {
            "description": "成功"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2026-09-08"
            }
          }
        ]
      }
    },
    "/api/v1/foods": {
      "get": {
        "summary": "名称搜索和分页列表",
        "operationId": "get_api_v1_foods",
        "tags": [
          "foods"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodList"
                },
                "example": {
                  "items": [
                    {
                      "id": 1,
                      "created_at": "2026-09-08T04:00:00Z",
                      "updated_at": "2026-09-08T04:00:00Z",
                      "name": "牛奶",
                      "basis_unit": "ml",
                      "nutrition_per_100": {
                        "energy_kcal": 60,
                        "protein_g": 3,
                        "carbohydrate_g": 5,
                        "fat_g": 3,
                        "sodium_mg": null,
                        "calcium_mg": null,
                        "vitamin_c_mg": null,
                        "vitamin_d_ug": null
                      }
                    }
                  ],
                  "page": 1,
                  "page_size": 20,
                  "total": 1
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1,
              "maximum": 1000000
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ]
      },
      "post": {
        "summary": "创建Food",
        "operationId": "post_api_v1_foods",
        "tags": [
          "foods"
        ],
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Food"
                },
                "example": {
                  "id": 1,
                  "created_at": "2026-09-08T04:00:00Z",
                  "updated_at": "2026-09-08T04:00:00Z",
                  "name": "牛奶",
                  "basis_unit": "ml",
                  "nutrition_per_100": {
                    "energy_kcal": 60,
                    "protein_g": 3,
                    "carbohydrate_g": 5,
                    "fat_g": 3,
                    "sodium_mg": null,
                    "calcium_mg": null,
                    "vitamin_c_mg": null,
                    "vitamin_d_ug": null
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodInput"
              },
              "example": {
                "name": "牛奶",
                "basis_unit": "ml",
                "nutrition_per_100": {
                  "energy_kcal": 60,
                  "protein_g": 3,
                  "carbohydrate_g": 5,
                  "fat_g": 3,
                  "sodium_mg": null
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/foods/{id}": {
      "get": {
        "summary": "查询Food",
        "operationId": "get_api_v1_foods_id",
        "tags": [
          "foods"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Food"
                },
                "example": {
                  "id": 1,
                  "created_at": "2026-09-08T04:00:00Z",
                  "updated_at": "2026-09-08T04:00:00Z",
                  "name": "牛奶",
                  "basis_unit": "ml",
                  "nutrition_per_100": {
                    "energy_kcal": 60,
                    "protein_g": 3,
                    "carbohydrate_g": 5,
                    "fat_g": 3,
                    "sodium_mg": null,
                    "calcium_mg": null,
                    "vitamin_c_mg": null,
                    "vitamin_d_ug": null
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ]
      },
      "patch": {
        "summary": "更新Food",
        "operationId": "patch_api_v1_foods_id",
        "tags": [
          "foods"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Food"
                },
                "example": {
                  "id": 1,
                  "created_at": "2026-09-08T04:00:00Z",
                  "updated_at": "2026-09-08T04:00:00Z",
                  "name": "牛奶",
                  "basis_unit": "ml",
                  "nutrition_per_100": {
                    "energy_kcal": 60,
                    "protein_g": 3,
                    "carbohydrate_g": 5,
                    "fat_g": 3,
                    "sodium_mg": null,
                    "calcium_mg": null,
                    "vitamin_c_mg": null,
                    "vitamin_d_ug": null
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodPatch"
              },
              "example": {
                "nutrition_per_100": {
                  "protein_g": 3.2,
                  "sodium_mg": null
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recipes": {
      "get": {
        "summary": "名称搜索和分页列表",
        "operationId": "get_api_v1_recipes",
        "tags": [
          "recipes"
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeList"
                },
                "example": {
                  "items": [
                    {
                      "id": 1,
                      "created_at": "2026-09-08T04:00:00Z",
                      "updated_at": "2026-09-08T04:00:00Z",
                      "name": "热牛奶",
                      "ingredients": [
                        {
                          "food_id": 1,
                          "name": "牛奶",
                          "basis_unit": "ml",
                          "quantity": 250,
                          "nutrition_per_100": {
                            "energy_kcal": 60,
                            "protein_g": 3,
                            "carbohydrate_g": 5,
                            "fat_g": 3,
                            "sodium_mg": null,
                            "calcium_mg": null,
                            "vitamin_c_mg": null,
                            "vitamin_d_ug": null
                          }
                        }
                      ],
                      "nutrition_totals": {
                        "energy_kcal": {
                          "known_total": 150.0,
                          "complete": true
                        },
                        "protein_g": {
                          "known_total": 7.5,
                          "complete": true
                        },
                        "carbohydrate_g": {
                          "known_total": 12.5,
                          "complete": true
                        },
                        "fat_g": {
                          "known_total": 7.5,
                          "complete": true
                        },
                        "sodium_mg": {
                          "known_total": null,
                          "complete": false
                        },
                        "calcium_mg": {
                          "known_total": null,
                          "complete": false
                        },
                        "vitamin_c_mg": {
                          "known_total": null,
                          "complete": false
                        },
                        "vitamin_d_ug": {
                          "known_total": null,
                          "complete": false
                        }
                      }
                    }
                  ],
                  "page": 1,
                  "page_size": 20,
                  "total": 1
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1,
              "maximum": 1000000
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ]
      },
      "post": {
        "summary": "创建Recipe",
        "operationId": "post_api_v1_recipes",
        "tags": [
          "recipes"
        ],
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipe"
                },
                "example": {
                  "id": 1,
                  "created_at": "2026-09-08T04:00:00Z",
                  "updated_at": "2026-09-08T04:00:00Z",
                  "name": "热牛奶",
                  "ingredients": [
                    {
                      "food_id": 1,
                      "name": "牛奶",
                      "basis_unit": "ml",
                      "quantity": 250,
                      "nutrition_per_100": {
                        "energy_kcal": 60,
                        "protein_g": 3,
                        "carbohydrate_g": 5,
                        "fat_g": 3,
                        "sodium_mg": null,
                        "calcium_mg": null,
                        "vitamin_c_mg": null,
                        "vitamin_d_ug": null
                      }
                    }
                  ],
                  "nutrition_totals": {
                    "energy_kcal": {
                      "known_total": 150.0,
                      "complete": true
                    },
                    "protein_g": {
                      "known_total": 7.5,
                      "complete": true
                    },
                    "carbohydrate_g": {
                      "known_total": 12.5,
                      "complete": true
                    },
                    "fat_g": {
                      "known_total": 7.5,
                      "complete": true
                    },
                    "sodium_mg": {
                      "known_total": null,
                      "complete": false
                    },
                    "calcium_mg": {
                      "known_total": null,
                      "complete": false
                    },
                    "vitamin_c_mg": {
                      "known_total": null,
                      "complete": false
                    },
                    "vitamin_d_ug": {
                      "known_total": null,
                      "complete": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeInput"
              },
              "example": {
                "name": "燕麦牛奶",
                "ingredients": [
                  {
                    "food_id": 1,
                    "quantity": 0.25,
                    "unit": "L"
                  },
                  {
                    "food_id": 2,
                    "quantity": 50,
                    "unit": "g"
                  }
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GoalInput": {
        "type": "object",
        "properties": {
          "energy_kcal": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "protein_percent": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 100
          },
          "carbohydrate_percent": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 100
          },
          "fat_percent": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 100
          }
        },
        "required": [
          "energy_kcal",
          "protein_percent",
          "carbohydrate_percent",
          "fat_percent"
        ],
        "additionalProperties": false,
        "description": "供能百分比合计为 100，克数按蛋白质/碳水 4 kcal/g、脂肪 9 kcal/g 推导。"
      },
...
      "Goal": {
        "type": "object",
        "properties": {
          "energy_kcal": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "protein_percent": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 100
          },
          "carbohydrate_percent": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 100
          },
          "fat_percent": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 100
          },
          "targets": {
            "$ref": "#/components/schemas/Targets"
          },
          "effective_on": {
            "type": "string",
            "format": "date",
            "example": "2026-09-08"
          }
        },
        "required": [
          "energy_kcal",
          "protein_percent",
          "carbohydrate_percent",
          "fat_percent",
          "targets"
        ],
        "additionalProperties": false
      },
      "ResolvedGoal": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "example": "2026-09-08"
          },
          "source": {
            "type": "string",
            "enum": [
              "default",
              "override",
              "none"
            ]
          },
          "goal": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Goal"
              }
            ],
            "nullable": true
          }
        },
        "required": [
          "date",
          "source",
          "goal"
        ],
        "additionalProperties": false
      },
      "Food": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "basis_unit": {
            "type": "string",
            "enum": [
              "g",
              "ml"
            ]
          },
          "nutrition_per_100": {
            "$ref": "#/components/schemas/Nutrition"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "name",
          "basis_unit",
          "nutrition_per_100"
        ],
        "additionalProperties": false
      },
      "FoodInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "basis_unit": {
            "type": "string",
            "enum": [
              "g",
              "ml"
            ]
          },
          "nutrition_per_100": {
            "$ref": "#/components/schemas/NutritionInput"
          }
        },
        "required": [
          "name",
          "basis_unit",
          "nutrition_per_100"
        ],
        "additionalProperties": false
      },
      "FoodPatch": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "nutrition_per_100": {
            "$ref": "#/components/schemas/NutritionPatch"
          }
        },
        "required": [],
        "additionalProperties": false,
        "description": "只更新提供的字段；营养对象逐字段合并，可选营养显式 null 清空，热量不可为空；basis_unit 创建后不可修改。"
      },
...
      "RecipeInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IngredientInput"
            },
            "minItems": 1
          }
        },
        "required": [
          "name",
          "ingredients"
        ],
        "additionalProperties": false
      },
...
      "FoodList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Food"
            }
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "Error400": {
        "description": "JSON 格式、字段或类型错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "bad_request",
                "message": "JSON 格式、字段或类型错误",
                "fields": {}
              }
            }
          }
        }
      },
      "Error404": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "not_found",
                "message": "资源不存在",
                "fields": {}
              }
            }
          }
        }
      },
      "Error422": {
        "description": "业务参数不合法",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "validation_error",
                "message": "业务参数不合法",
                "fields": {
                  "quantity": "数量必须大于 0"
                }
              }
            }
          }
        }
      },
      "Error500": {
        "description": "内部错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "internal_error",
                "message": "内部错误",
                "fields": {}
              }
            }
          }
        }
      }
    }
  }
}